Add username change + courier order adjust + nearest store; fix quantity float migration
- admin/business/platform auth: allow changing login (username) alongside password
- courier: PATCH /orders/{id}/adjust to edit item quantities before delivery starts
- catalog: GET /restaurants/nearest picks closest active store by coordinates
- order_items.quantity is now Float (fractional units like kg); initdb.py migrates
  the existing INTEGER column, courier adjust totals round instead of truncate, and
  the top-products report no longer truncates fractional quantities to int
- settings pages: surface "login band" errors distinctly from password errors

+++ backend/app/api/routes/admin.py
@@ -223,7 +223,7 @@ def _top_products(db: Session, rids: list[int], limit: int = 20) -> list[TopProd
     return [
         TopProduct(
             product_id=pid, name_uz=name, image_url=img,
-            quantity=int(qty), revenue=int(rev), profit=int(prof),
+            quantity=float(qty), revenue=int(rev), profit=int(prof),
         )
         for pid, name, img, qty, rev, prof in rows
     ]
